requirements.txt
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pynput==1.7.7
PyQt6==6.7.1
PyQt6-Fluent-Widgets==1.7.4
loguru==0.7.3
pywin32==308